Extract inferece model stats into its own file
This commit extracts the inner class `ModelStats` of
`InferenceFeatureSetUsage` into its own file.

There are 2 reasons for doing this:

  - Allows testing of the `InferenceFeatureSetUsage` class
  - Introduces a `usage` package under `inference` that can be
  used to accommodate additional usage data.
