+++ x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/inference/InferenceFeatureSetUsageTests.java
@@ -7,35 +7,35 @@
 
 package org.elasticsearch.xpack.core.inference;
 
-import com.carrotsearch.randomizedtesting.generators.RandomStrings;
-
 import org.elasticsearch.common.io.stream.Writeable;
-import org.elasticsearch.inference.TaskType;
 import org.elasticsearch.test.AbstractWireSerializingTestCase;
+import org.elasticsearch.xpack.core.inference.usage.ModelStats;
+import org.elasticsearch.xpack.core.inference.usage.ModelStatsTests;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 
-public class InferenceFeatureSetUsageTests extends AbstractWireSerializingTestCase<InferenceFeatureSetUsage.ModelStats> {
+public class InferenceFeatureSetUsageTests extends AbstractWireSerializingTestCase<InferenceFeatureSetUsage> {
 
     @Override
-    protected Writeable.Reader<InferenceFeatureSetUsage.ModelStats> instanceReader() {
-        return InferenceFeatureSetUsage.ModelStats::new;
+    protected Writeable.Reader<InferenceFeatureSetUsage> instanceReader() {
+        return InferenceFeatureSetUsage::new;
     }
 
     @Override
-    protected InferenceFeatureSetUsage.ModelStats createTestInstance() {
-        RandomStrings.randomAsciiLettersOfLength(random(), 10);
-        return new InferenceFeatureSetUsage.ModelStats(
-            randomIdentifier(),
-            TaskType.values()[randomInt(TaskType.values().length - 1)],
-            randomInt(10)
-        );
+    protected InferenceFeatureSetUsage createTestInstance() {
+        return new InferenceFeatureSetUsage(randomList(10, ModelStatsTests::createRandomInstance));
     }
 
     @Override
-    protected InferenceFeatureSetUsage.ModelStats mutateInstance(InferenceFeatureSetUsage.ModelStats modelStats) throws IOException {
-        InferenceFeatureSetUsage.ModelStats newModelStats = new InferenceFeatureSetUsage.ModelStats(modelStats);
-        newModelStats.add();
-        return newModelStats;
+    protected InferenceFeatureSetUsage mutateInstance(InferenceFeatureSetUsage instance) throws IOException {
+        List<ModelStats> mutatedModelStats = new ArrayList<>(instance.modelStats());
+        if (mutatedModelStats.isEmpty()) {
+            mutatedModelStats.add(ModelStatsTests.createRandomInstance());
+        } else {
+            mutatedModelStats.remove(randomIntBetween(0, mutatedModelStats.size() - 1));
+        }
+        return new InferenceFeatureSetUsage(mutatedModelStats);
     }
 }
